chore(pipeline): refatctor the pipeline directory
