+++ src/query/pipeline/sinks/Cargo.toml
@@ -1,7 +1,10 @@
 [package]
 name = "common-pipeline-sinks"
-version = "0.1.0"
-edition = "2021"
+version = { workspace = true }
+authors = { workspace = true }
+license = { workspace = true }
+publish = { workspace = true }
+edition = { workspace = true }
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [lib]
