+++ src/query/pipeline/sinks/src/empty_sink.rs
