+++ src/query/pipeline/sinks/src/lib.rs
@@ -14,4 +14,18 @@
 
 #![feature(type_alias_impl_trait)]
 
-pub mod processors;
+mod async_sink;
+mod context_sink;
+mod empty_sink;
+mod sync_sink;
+mod sync_sink_sender;
+mod union_receive_sink;
+
+pub use async_sink::AsyncSink;
+pub use async_sink::AsyncSinker;
+pub use context_sink::ContextSink;
+pub use empty_sink::EmptySink;
+pub use sync_sink::Sink;
+pub use sync_sink::Sinker;
+pub use sync_sink_sender::SyncSenderSink;
+pub use union_receive_sink::UnionReceiveSink;
