+++ src/query/pipeline/sinks/src/sync_sink.rs
