+++ src/query/service/tests/it/pipelines/executor/executor_graph.rs
@@ -22,11 +22,11 @@ use common_exception::Result;
 use common_expression::DataBlock;
 use common_pipeline_core::pipe::Pipe;
 use common_pipeline_core::pipe::PipeItem;
+use common_pipeline_sinks::SyncSenderSink;
 use common_pipeline_sources::processors::sources::SyncReceiverSource;
 use databend_query::pipelines::executor::RunningGraph;
 use databend_query::pipelines::processors::port::InputPort;
 use databend_query::pipelines::processors::port::OutputPort;
-use databend_query::pipelines::processors::SyncSenderSink;
 use databend_query::pipelines::processors::TransformDummy;
 use databend_query::pipelines::Pipeline;
 use databend_query::sessions::QueryContext;
