+++ src/query/storages/null/src/null_table.rs
@@ -30,7 +30,7 @@ use common_meta_app::schema::TableInfo;
 use common_pipeline_core::processors::port::OutputPort;
 use common_pipeline_core::processors::processor::ProcessorPtr;
 use common_pipeline_core::Pipeline;
-use common_pipeline_sinks::processors::sinks::EmptySink;
+use common_pipeline_sinks::EmptySink;
 use common_pipeline_sources::processors::sources::SyncSource;
 use common_pipeline_sources::processors::sources::SyncSourcer;
